Consolidate the slot fill ordering logic
