+++ plugins/woocommerce-admin/client/utils/index.js
@@ -1,4 +1,5 @@
 export * from './plugins';
+export * from './slot-fill-ordering';
 
 /**
  * Get the URL params.
